requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
docopt==0.6.2
engineering-notation==0.6.0
et-xmlfile==1.1.0
exceptiongroup==1.0.4
execnet==1.9.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
pipreqs==0.4.11
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
pytest==7.1.2
pytest-forked==1.4.0
pytest-html-reporter==0.2.9
pytest-xdist==3.0.2
PyYAML==6.0
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.10
wsproto==1.1.0
yarg==0.1.9
